7주차 완료 및 예외처리 클래스 정의

+++ src/main/java/com/example/bcsd/models/Member.java
@@ -2,23 +2,23 @@ package com.example.bcsd.models;
 
 public class Member {
 
-    private Long id;
+    private int id;
     private String name;
     private String email;
     private String password;
 
-    public Member(Long id, String name, String email, String password) {
+    public Member(int id, String name, String email, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.password = password;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
